fixed clear score issue

+++ Assets/Leaderboard.js
@@ -6,45 +6,42 @@ let total = localStorage.getItem("totalScore");
 let array = [];
 let totalInput = nameInput + " " + total;
 
-
-
 function renderScores() {
-  if (array !== null) {
+  if (total !== null) {
     for (var i = 0; i < array.length; i++) {
       var createLi = document.createElement("li");
       createLi.textContent = array[i];
       list.appendChild(createLi);
     }
-  } 
+  } else {
+    localStorage.clear();
+  }
 }
 
+//store leaderboard data
 function storeScores() {
   localStorage.setItem("final", JSON.stringify(array));
+  
 }
 
+//pulls past data into the leaderboard
 function getScores() {
   let storeArray = JSON.parse(localStorage.getItem("final"));
 
   if (storeArray !== null) {
     array = storeArray;
-    
-}
-    array.push(totalInput);
+  }
+  //adds new array to existing one
+  array.push(totalInput);
 
   renderScores();
   storeScores();
 }
-getScores();
-
-function ClearBoard() {
-  window.localStorage.removeItem("final");
-  // window.location.reload;
-  localStorage.clear();
-  
+if (total !== null) {
+  getScores();
 }
-document.getElementById("erase").onclick = ClearBoard();
-// erase.addEventListener("click", function () {
- 
-//   localStorage.clear();
-//   location.reload();
-// });
+//erases all stored data and shows empty leaderboard 
+erase.addEventListener("click", function () {
+  localStorage.clear();
+  location.reload();
+});
